feat about getViewMode
