+++ src/components/MainPageHero/PreviewCategories/MainGallery.jsx
@@ -1,51 +1,26 @@
 import React, { useEffect, useState } from "react";
-import axios from "axios";
-import NoImage from "../../../images/NoImage/NoImageSmall.png";
+import { GalleryUl, MainGalleryH2, ButtonWrapper } from "./MainGallery.styled";
+import { PreviewCategories } from "./PreviewCategories";
+import { fetchCategories } from "../../../api/mainRecipesApi";
+import { getRecipesLimit , getViewMode} from "../../../api/viewModeUtils";
 import { ButtonSeeAll } from "../../Buttons/GalleryButtonSeeAll";
-import {
-  GalleryUl,
-  GalleryLi,
-  RecipeImg,
-  RecipeDescription,
-  MainGalleryH2,
-  ElementWrapper,
-  ButtonWrapper
-
-} from "./MainGallery.styled";
-
-export const baseAxiosURL = "http://localhost:5001/api";
-
-axios.defaults.baseURL = baseAxiosURL;
 
 export const MainGallery = () => {
   const [categories, setCategories] = useState([]);
-  const [viewMode, setViewMode] = useState("desktop");
+  const [viewMode, setViewMode] = useState(getViewMode());
 
   useEffect(() => {
-    const fetchCategories = async () => {
+    const fetchData = async () => {
       try {
-        const response = await axios.get("/recipes/main-page");
-
-        const categoriesArray = Object.values(response.data.recipesMainPage);
-        setCategories(categoriesArray);
+        const data = await fetchCategories();
+        setCategories(data);
       } catch (error) {
         console.error("Error fetching categories:", error);
       }
     };
 
-    fetchCategories();
+    fetchData();
   }, []);
-
-  const getViewMode = () => {
-    const width = window.innerWidth;
-    if (width >= 1280) {
-      return "desktop";
-    } else if (width >= 768) {
-      return "tablet";
-    } else {
-      return "mobile";
-    }
-  };
 
   useEffect(() => {
     const handleResize = () => {
@@ -57,47 +32,18 @@ export const MainGallery = () => {
     };
   }, []);
 
-  const renderRecipes = (recipes, limit) => {
-    return recipes.slice(0, limit).map((recipe, _id) => (
-      <ElementWrapper key={_id}>
 
-        <RecipeImg
-          src={recipe.thumb ? recipe.thumb : NoImage}
-          loading="lazy"
-          alt={recipe.title}
-        />
-        <RecipeDescription>
-          <p>{recipe.title}</p>
-        </RecipeDescription>
-      </ElementWrapper>
-    ));
-  };
-
-  const getRecipesLimit = () => {
-    switch (viewMode) {
-      case "mobile":
-        return 1;
-      case "tablet":
-        return 2;
-      case "desktop":
-        return categories.length;
-      default:
-        return 1;
-    }
-  };
 
   return (
     <GalleryUl>
-      {categories.map((categoryRecipes, _id) => (
-        <div key={_id}>
+      {categories.map((categoryRecipes, index) => (
+        <div key={index}>
           <MainGalleryH2>{categoryRecipes[0].category}</MainGalleryH2>
-          <GalleryLi key={_id}>
-            {renderRecipes(categoryRecipes, getRecipesLimit())}
-          </GalleryLi>
-
-        <ButtonWrapper><ButtonSeeAll text="See All"/></ButtonWrapper>
+          <PreviewCategories categoryRecipes={categoryRecipes} viewMode={viewMode} />
+          <ButtonWrapper>
+            <ButtonSeeAll text="See All" />
+          </ButtonWrapper>
         </div>
- 
       ))}
     </GalleryUl>
   );
