+++ src/components/MainPageHero/PreviewCategories/PreviewCategories.jsx
@@ -1,10 +1,35 @@
 import React from "react";
-import { MainGallery } from "./MainGallery";
+import { getRecipesLimit } from "../../../api/viewModeUtils";
+import {
+  GalleryLi,
+  ElementWrapper,
+  RecipeImg,
+  RecipeDescription,
+} from "./MainGallery.styled";
+import NoImage from "../../../images/NoImage/NoImageSmall.png";
 
-export function PreviewCategories (){
+export const PreviewCategories = ({ categoryRecipes, viewMode }) => {
+  const renderRecipes = (recipes, limit) => {
+    return recipes.slice(0, limit).map((recipe, index) => (
+      <ElementWrapper key={index}>
+        <RecipeImg
+          src={recipe.thumb ? recipe.thumb : NoImage}
+          loading="lazy"
+          alt={recipe.title}
+        />
+        <RecipeDescription>
+          <p>{recipe.title}</p>
+        </RecipeDescription>
+      </ElementWrapper>
+    ));
+  };
 
-    return(
-<MainGallery/>
-
-    )
-}
+  return (
+    <GalleryLi>
+      {renderRecipes(
+        categoryRecipes,
+        getRecipesLimit(viewMode, categoryRecipes)
+      )}
+    </GalleryLi>
+  );
+};
